Use SSTables for key lookups
Extend the search logic to consult the SSTables if the key does not
exist in the memtables. The SSTables do not have an index currently.
The search is linear.

Also, extend the KVStore constructor to discover the SSTables on
startup. The SSTable files have a positive integer as a suffix, which is
how the KVStore sorts them in chronological order.

Refs #4.

Signed-off-by: Nikos Dragazis <dragazisn@gmail.com>