Make key deletions persistent
Modify the MemTable business logic to store deletion markers
("tombstones") instead of deleting keys from the in-memory map.

Extend the SSTable as well to store the deletion markers on disk.

Finally, simplify the DELETE API to not return the previous value. This
would potentially require to search on disk, which would hinder its
performance.

Refs #3.
Refs #4.

Signed-off-by: Nikos Dragazis <dragazisn@gmail.com>