Make sure to clean up the HTTP server in case of exceptions
Starting the server could throw an exception, in which case the server
will never be stopped.

Add a try..catch block to ensure we stop the server in all cases. The
code is based on the Seastar tutorial page.

`seastar::defer()` would be a cleaner option, but it does not work with
coroutines.

Refs #5.

Signed-off-by: Nikos Dragazis <dragazisn@gmail.com>