Extract KVStore initialization/finalization logic into coroutines
A class constructor/destructor cannot be a coroutine or any other
asynchronous function.

Introduce `KVStore::start()/`KVStore::stop()` member functions to handle
the initialization and finalization asynchronously via Seastar.

Not sure if this is a best practice, as it breaks RAII, but I have seen
this pattern in ScyllaDB's codebase as well.

Refs #5.

Signed-off-by: Nikos Dragazis <dragazisn@gmail.com>