Extend GET /keys/ to report keys from all shards
Extend the function handler for the GET /keys/ endpoint to obtain all
keys from all shards, merge them, and return then as a json object.

Since data are sharded based on hash ranges, keys across shards are not
ordered, so we need to merge them manually on a key-by-key basis.

Closes #2.
Refs #6.

Signed-off-by: Nikos Dragazis <dragazisn@gmail.com>